Move Settings components somewhere sane
