+++ src/pages/settings/Import.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import type { DatabaseSchema } from "../../model/DatabaseSchema";
 import type { Entry } from "@zip.js/zip.js";
-import ActionButton from "../buttons/ActionButton.vue";
-import FileInput from "../attachments/FileInput.vue";
+import ActionButton from "../../components/buttons/ActionButton.vue";
+import FileInput from "../../components/attachments/FileInput.vue";
 import ImportProcessModal from "./ImportProcessModal.vue";
 import { BlobReader, TextWriter, ZipReader } from "@zip.js/zip.js";
 import { create } from "superstruct";
