+++ src/pages/settings/Settings.vue
